Move benchmarking mock into its own file
