+++ frame/evm/src/lib.rs
@@ -58,6 +58,7 @@ pub mod runner;
 
 #[cfg(any(test, feature = "runtime-benchmarks"))]
 pub mod benchmarks;
+mod mock;
 
 pub use crate::runner::Runner;
 pub use fp_evm::{
